natlas-server/requirements.txt
alembic==1.0.8
blinker==1.4
Click==7.0
dnspython==1.16.0
elasticsearch==6.3.1
email-validator==1.0.3
Flask==1.0.2
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.4.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.8
MarkupSafe==1.1.1
mpmath==1.1.0
netaddr==0.7.19
PyJWT==1.7.1
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
python-libnmap==0.7.0
six==1.12.0
SQLAlchemy==1.3.2
sympy==1.4
urllib3==1.24.2
Werkzeug==0.15.2
WTForms==2.2.1
